emsmdb: adjust fx_divisor to return size_t
For synchronization streams over 281470681743360 bytes (255.996 TB)
in length, fx_divisor would return the wrong value, which could
probably lead to funky progress bars in Outlook.

+++ exch/emsmdb/common_util.h
@@ -124,8 +124,8 @@ extern unsigned int g_max_rcpt, g_max_message, g_max_mail_len;
 extern unsigned int g_max_rule_len, g_max_extrule_len;
 extern char g_emsmdb_org_name[256];
 
-static inline uint32_t fx_divisor(uint64_t total)
+static inline size_t fx_divisor(size_t total)
 {
-	uint32_t r = total / 0xFFFF;
+	size_t r = total / 0xFFFF;
 	return r > 0 ? r : 1;
 }
